move and cleanup Svelte components
