+++ code/renderers/svelte/src/components/SlotDecorator.svelte
