--- code/renderers/svelte/templates/HOC.svelte
@@ -1,7 +1,0 @@
-<script>
-  export let storyFn;
-
-  let { Component: component, props } = storyFn();
-</script>
-
-<svelte:component this={component} {...props}/>
